refactor: move HostPort out of NetworkListenerConfiguration
