+++ carapace-server/src/main/java/org/carapaceproxy/core/RuntimeServerConfiguration.java
@@ -41,6 +41,7 @@ import lombok.Data;
 import org.carapaceproxy.configstore.ConfigurationStore;
 import org.carapaceproxy.server.config.ConfigurationNotValidException;
 import org.carapaceproxy.server.config.ConnectionPoolConfiguration;
+import org.carapaceproxy.server.config.HostPort;
 import org.carapaceproxy.server.config.NetworkListenerConfiguration;
 import org.carapaceproxy.server.config.RequestFilterConfiguration;
 import org.carapaceproxy.server.config.SSLCertificateConfiguration;
@@ -480,7 +481,7 @@ public class RuntimeServerConfiguration {
         return requestFilters;
     }
 
-    NetworkListenerConfiguration getListener(NetworkListenerConfiguration.HostPort hostPort) {
+    NetworkListenerConfiguration getListener(HostPort hostPort) {
         return listeners
                 .stream()
                 .filter(s -> s.getHost().equalsIgnoreCase(hostPort.host()) && s.getPort() == hostPort.port())
