+++ carapace-server/src/main/java/org/carapaceproxy/server/config/HostPort.java
@@ -1,0 +1,4 @@
+package org.carapaceproxy.server.config;
+
+public record HostPort(String host, int port) {
+}
