+++ carapace-server/src/main/java/org/carapaceproxy/server/config/NetworkListenerConfiguration.java
@@ -168,7 +168,4 @@ public class NetworkListenerConfiguration {
         }
         return Set.of(HTTP11.name(), H2C.name());
     }
-
-    public record HostPort(String host, int port) {
-    }
 }
