+++ carapace-server/src/test/java/org/carapaceproxy/listeners/ListenerConfigurationTest.java
@@ -8,7 +8,7 @@ import org.carapaceproxy.configstore.PropertiesConfigurationStore;
 import org.carapaceproxy.core.HttpProxyServer;
 import org.carapaceproxy.core.Listeners;
 import org.carapaceproxy.server.config.ConfigurationChangeInProgressException;
-import org.carapaceproxy.server.config.NetworkListenerConfiguration;
+import org.carapaceproxy.server.config.HostPort;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
@@ -32,12 +32,10 @@ public class ListenerConfigurationTest {
             }
             server.start();
 
-            NetworkListenerConfiguration.HostPort listenerKey =
-                    new NetworkListenerConfiguration.HostPort("localhost", 8080);
+            HostPort listenerKey = new HostPort("localhost", 8080);
 
             {
-                Map<NetworkListenerConfiguration.HostPort, Listeners.ListeningChannel> listeners =
-                        server.getListeners().getListeningChannels();
+                Map<HostPort, Listeners.ListeningChannel> listeners = server.getListeners().getListeningChannels();
 
                 //check default configuration
                 assertEquals(true, listeners.get(listenerKey).getConfig().isKeepAlive());
@@ -57,8 +55,7 @@ public class ListenerConfigurationTest {
 
                 reloadConfiguration(configuration, server);
 
-                Map<NetworkListenerConfiguration.HostPort, Listeners.ListeningChannel> listeners =
-                        server.getListeners().getListeningChannels();
+                Map<HostPort, Listeners.ListeningChannel> listeners = server.getListeners().getListeningChannels();
 
                 assertEquals(1, listeners.size());
                 assertEquals(false, listeners.get(listenerKey).getConfig().isKeepAlive());
@@ -79,8 +76,7 @@ public class ListenerConfigurationTest {
                 configuration.put("listener.1.enabled", "true");
                 reloadConfiguration(configuration, server);
 
-                Map<NetworkListenerConfiguration.HostPort, Listeners.ListeningChannel> listeners =
-                        server.getListeners().getListeningChannels();
+                Map<HostPort, Listeners.ListeningChannel> listeners = server.getListeners().getListeningChannels();
 
                 assertEquals(true, listeners.get(listenerKey).getConfig().isKeepAlive());
                 assertEquals(10, listeners.get(listenerKey).getConfig().getSoBacklog());
